python/requirements.txt
amici==0.31.1
antimony==2.12.0.2
asttokens==3.0.0
cmake==4.0.0
cmake-build-extension==0.6.0
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
gitdb==4.0.12
GitPython==3.1.44
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
ninja==1.11.1.4
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-libsbml==5.20.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
setuptools==78.1.0
setuptools-scm==8.2.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.3
toposort==1.10
tornado==6.4.2
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
